+++ src/tests/sec/test_filter.py
@@ -1,0 +1,68 @@
+import stocktracer.filter as Filter
+from stocktracer.data.sec import (
+    ReportDate,
+)
+from datetime import date
+import pytest
+
+class TestFilter:
+    def test_getRequiredReports_default(self):
+        filter = Filter.SecFilter(
+            tags=["dummy"],
+            last_report=ReportDate(year=2022, quarter=4),
+        )
+
+        required_reports = filter.required_reports
+
+        # You might thing 1, but since companies file annual reports in different quarters,
+        # we have to look at all the quarters.
+        assert len(required_reports) == 5
+        assert required_reports[0] == ReportDate(year=2022, quarter=4)
+        assert required_reports[1] == ReportDate(year=2022, quarter=3)
+        assert required_reports[2] == ReportDate(year=2022, quarter=2)
+        assert required_reports[3] == ReportDate(year=2022, quarter=1)
+        assert required_reports[4] == ReportDate(year=2021, quarter=4)
+
+    def test_getRequiredReports_quarterly(self):
+        filter = Filter.SecFilter(
+            tags=["dummy"],
+            years=1,
+            last_report=ReportDate(year=2022, quarter=4),
+            only_annual=False,
+        )
+        required_reports = filter.required_reports
+
+        assert len(required_reports) == 5
+        assert required_reports[0] == ReportDate(year=2022, quarter=4)
+        assert required_reports[1] == ReportDate(year=2022, quarter=3)
+        assert required_reports[2] == ReportDate(year=2022, quarter=2)
+        assert required_reports[3] == ReportDate(year=2022, quarter=1)
+        assert required_reports[4] == ReportDate(year=2021, quarter=4)
+
+
+
+def test_reportDate():
+    rd = ReportDate(2023, 1)
+    assert rd.year == date.today().year
+    assert rd.quarter == 1
+
+    try:
+        rd = ReportDate(date.today().year + 1, 1)
+        pytest.fail("should throw and exception")
+    except ValueError as ex:
+        pass
+
+    try:
+        rd = ReportDate(date.today().year, 0)
+        pytest.fail("should throw and exception")
+    except ValueError as ex:
+        pass
+
+    rd = ReportDate(date.today().year, 4)
+    try:
+        rd = ReportDate(date.today().year, 5)
+        pytest.fail("should throw and exception")
+    except ValueError as ex:
+        pass
+
+
